+++ packages/terra-i18n/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added more information to upgrade guide around Jest testing
 
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
 4.6.0 - (June 18, 2019)
 ------------------
 ### Changed